+++ src/main/java/com/powsybl/openloadflow/network/AbstractElement.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.openloadflow.network;
+
+import java.util.Objects;
+
+/**
+ * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ */
+public abstract class AbstractElement {
+
+    protected final LfNetwork network;
+
+    protected int num = -1;
+
+    protected AbstractElement(LfNetwork network) {
+        this.network = Objects.requireNonNull(network);
+    }
+
+    public int getNum() {
+        return num;
+    }
+
+    public void setNum(int num) {
+        this.num = num;
+    }
+
+    public LfNetwork getNetwork() {
+        return network;
+    }
+}
