+++ src/main/java/com/powsybl/openloadflow/network/AbstractLfBranch.java
@@ -17,13 +17,9 @@ import java.util.Objects;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public abstract class AbstractLfBranch implements LfBranch {
+public abstract class AbstractLfBranch extends AbstractElement implements LfBranch {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractLfBranch.class);
-
-    private int num = -1;
-
-    private final LfNetwork network;
 
     private final LfBus bus1;
 
@@ -36,25 +32,10 @@ public abstract class AbstractLfBranch implements LfBranch {
     protected DiscreteVoltageControl discreteVoltageControl;
 
     protected AbstractLfBranch(LfNetwork network, LfBus bus1, LfBus bus2, PiModel piModel) {
-        this.network = Objects.requireNonNull(network);
+        super(network);
         this.bus1 = bus1;
         this.bus2 = bus2;
         this.piModel = Objects.requireNonNull(piModel);
-    }
-
-    @Override
-    public int getNum() {
-        return num;
-    }
-
-    @Override
-    public void setNum(int num) {
-        this.num = num;
-    }
-
-    @Override
-    public LfNetwork getNetwork() {
-        return network;
     }
 
     @Override
