+++ src/main/java/com/powsybl/openloadflow/network/impl/LfShuntImpl.java
@@ -7,6 +7,8 @@
 package com.powsybl.openloadflow.network.impl;
 
 import com.powsybl.iidm.network.ShuntCompensator;
+import com.powsybl.openloadflow.network.AbstractElement;
+import com.powsybl.openloadflow.network.LfNetwork;
 import com.powsybl.openloadflow.network.LfShunt;
 import com.powsybl.openloadflow.network.PerUnit;
 import com.powsybl.openloadflow.util.Evaluable;
@@ -18,17 +20,16 @@ import static com.powsybl.openloadflow.util.EvaluableConstants.NAN;
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class LfShuntImpl implements LfShunt {
+public class LfShuntImpl extends AbstractElement implements LfShunt {
 
     private final ShuntCompensator shuntCompensator;
-
-    private int num = -1;
 
     private final double b;
 
     private Evaluable q = NAN;
 
-    public LfShuntImpl(ShuntCompensator shuntCompensator) {
+    public LfShuntImpl(ShuntCompensator shuntCompensator, LfNetwork network) {
+        super(network);
         this.shuntCompensator = Objects.requireNonNull(shuntCompensator);
         double nominalV = shuntCompensator.getTerminal().getVoltageLevel().getNominalV();
         double zb = nominalV * nominalV / PerUnit.SB;
@@ -38,14 +39,6 @@ public class LfShuntImpl implements LfShunt {
     @Override
     public String getId() {
         return shuntCompensator.getId();
-    }
-
-    public int getNum() {
-        return num;
-    }
-
-    public void setNum(int num) {
-        this.num = num;
     }
 
     @Override
